The dataset of worldwide mafic rocks
We downloaded all the geochemical data for the igneous rocks from the EarthChem database, including the whole-rock major and trace element concentrations, ages, and geographic coordinates. In consideration of their distinct geochemical characteristics we used only arc and continental mafic rocks in this study. A final dataset containing 58,674 whole-rock geochemical data of mafic igneous rocks covering 3.5-0 Ga was prepared for the statistical analysis.